remove failure crate in favor of std::error::Error
With the stabilization of std::error::Error::source(), the standard
Error trait provides everything that is provided by the Fail trait. All
failure was really being used for was to provide derivation of the
Display trait and a with_context call in one place in the library, which
was easy enough to provide for the standard Error trait.